MetaCPAN::Web::Controller::Search: don't implicitly append " 1" to IFL queries
This code was conflating the boolean test for checking if we had the
"q" or "lucky" paramaters with what we ended up searching for, so
e.g. searching for "XML" under IFL would implicitly search for "XML
1", which would return "Example-1", instead of "XML".

Change the code to only use "lucky" as a boolean parameter, not as a
part of the search query.