First stab at reports for CPAN-API/cpan-api#253
I'll let people fight over nomenclature. Meanwhile here's some first
stab at the report. :-)

I'm still not sure that per-module makes more
sense than per-distribution, but just for the sake of having something
running, I've assuming that the data is per module. If that change,
the logic will be easy enough to move around.

I'm not terribly happy about the name of the keys, they'll probably
have to be changed.

I've also sorted the recommendations per total activity, and report both
the recommendations in favor of alternatives and against in the same
list, which seems to me the most succinct way to present the information.