Merge Catalyst and Catalyst::Runtime dependencies
Catalyst and Catalyst::Runtime both come from a single distribution
(Catalyst-Runtime) and share a version number, so it isn't necessary to
specify both of them -- and it's confusing to specific a different number
for each.

Put the version dependency on Catalyst::Runtime, since that's the module
that MetaCPAN::Web actually tries to load a specific version of.