Stop trimming HTML in highlighter
Issues #657 and #626 both mention a problem of stripping leading whitespace on the
first line of a verbatim block. This change fixes that.

I do not know this code well enough to know if this will have other adverse effects

My rationale is that at this point the sh code has already stripped empty lines at
beginning and end, and unindented the code, so I would assume any left over white-
space to be important.