Include dev and backpan releases in cpantesters link
The CPAN Testers page has filters for dev releases and backpan releases.  By default, both are excluded.  Even if a dist version is listed explicitly in the link, it won't be selected if it is excluded by the filters.  Set the filters to include backpan and dev releases, so the links will work for all dist versions.