Revert "add rel=author links to Google+ to pod and release pages in order to provide Author information to Google"
This reverts commit 7c8b1d6c0d7b73db2402c7a3d37072c7129d2076.

I'm reverting this commit because it's a bit controversial.  The
following problems:

1) We already have the "plus" icon on author pages, so this duplicates
the image.

2) There has been some concern that this places Google in a more
prominent position than other search engines for no apparent gain.

3) It's not clear (to me) what the real difference between rel="me" and
rel="author" is.  Probably we should decide on how we want to handle
that and then do it without adding an extra image.

We do have the Coderwall endorse button under author images where
appropriate, but that button serves a different purpose and it also
relays some info about the author.

So, I think that with this particular commit, we'd just need some more
discussion as to whether it's really needed and how best to implement
it.