Change styling of identities buttons
The buttons used some custom styles to make them equal width and some strange pads.

Change to use `btn-block` from bootstrap to get block level buttons, which gives
them equal width, and remove `btn-large` as they are plenty easy to hit without
being large, and with there being so many of them in a row, this makes it a bit less
in your face.