Change the add/remove/check links to buttons
The motivation was "remove check", which looked like one link to me. Now it
is clearly two buttons, and one is red.

I opted to not add `btn-danger` to the remove buttons that are on distinctive lines,
as opposed to the profile where there are two buttons on each line.