Add higher resolution icons to favicon.ico
Adds a 32x32 and 48x48 icon to favicon.ico. It means the filesize increases slightly,
but it also means that it looks nicer in Firefox, Safari and Chrome tabs.

I also added apple-touch-icon (152x152, iOS7 retina version) if someone saves metacpan
to their "home screen" on mobile devices, this icon will be used.

The favicon.ico icons are transparent, the apple-touch-icon is on a white background.
iOS uses a black background for transparent parts and that looked pretty damn awful.