Use a terrible placeholder for users with missing gravatar URLs.
We won't skip users with missing icons because then the ++ number won't
add up.  Someone with a better icon can fix this or we leave it as is in
order to motivate us to clean up the cases where users blew away the
original URL and haven't been able to replace them.  I believe there is
an open issue for this.