Fix initial load of dep graph to display:none;
On initial load the dep graph ifram and div was overlaying everything, intercepting
clicks and scrolls (At least in firefox). After opening and dismissing it, it worked
and the only difference was that after dismissal, the style had changed to
`display: none;`. Applying `display: none` on initial load fixes the problem.