Change dependency graph link to a button. Fixes GH #1195.
Displaying the dependency graph modal isn't really
a "navagation" event.  So it is more appropriate to
implement this as a <button> rather than an <a> link.

However, Bootstrap has slightly different styling for
buttons, so I had to add a bit of CSS to make this particular
button look just like the neighboring links.

I also moved the modal <div> outside the <ul> since it wasn't
strictly valid, and it makes the markup easier to understand
when/if you need to look at the page source.