Add anchors to news, FAQ (and other about pages)
This is achieved by switching to MultiMarkdown, which has an option called
`heading_ids`, which auto-generated id tags for headings.

Secondly we re-use the anchors stuff from the POD pages, by changing to a shared
css class `anchors`, since the `pod` class changed text layout a lot.

Lastly, adjust some of the anchor CSS for the large H1 headings on about pages.

The CSS could probably be tweaked more, but it got messy real quick :(