clean up styles for sidebars
A lot of the rules were to override the nav styles to re-normalize
things.  The nav-list isn't part of bootstrap anymore, so just repurpose
it as our own sidebar class, and remove the nav class from the sidebars
using it.  This simplifies things greatly cleans up link sizes as well
as allowing arbitrary content easier.