clean up dispatch actions
This adds Args(0) to many of the endpoints to prevent them from
accepting arbitrary URLs with a matching prefix.  It also removes some
chained dispatching where the chaining wasn't taken advantage of at all.
Some of the subs were renamed to match their URL path, since there was
no need for them to be different.