Moved Elasticsearch::Adapter (single_valued_arrayref_to_scalar)
There is no real need for this to be in a Role, the functionality
does not depend on the object at all and it also makes it hard
to reuse by tests code.

This change moves it into an impotable function we can use
where needed.
There are few coming test fixes that depend on it.