Fix test t/controller/search/autocomplete.t
The way ES tokenizes the analyzed string in this search,
makes the search for 'moose">' match results for 'moose'
which is a better behavior than how it worked before.

In addition, the name prefix matching should not be
against the search string (which contains garbage) but
against the expected matching module name.