Short-circuit correctly in search_expanded
The total will always be the number of documents matched, not the number
of documents returned for the given page.  Instead, do what
search_collapsed does and check if we have any distributions.

Additionally, guard search_descriptions against an empty id list, which
produces invalid queries.

I found these while tracking down the cause of thousands of 500s on
/file/_search, ultimately being triggered by a web crawler.