Module search: Avoid requesting duplicate fields from ES
Requesting a field more than once produces the value more than once
which causes the new ES to produce an array value for the field
containing duplicate values.  This busts our collapsed search which
ended up making ref addresses unique instead of distribution names.
With the root cause removed, the map which papered over the symptoms is
not longer needed.