Fix the Bootstrap .clearfix class to actually fix clearing
I have no idea how or why clear: both turned into clear: none in commit
e11081f, which ostensibly upgraded to Bootstrap 3.1.1.  That release of
Bootstrap had clear: both in the &:after rule, so perhaps a spurious
local modification made it into the commit.

The upshot of this is that the keyboard shortcuts modal dialog isn't
busted any more.