Stop using floats poorly for the about/account sidebar
Instead, just use absolute positioning which doesn't play games with
other elements on the page.  (The problem was that some elements want to
clear floats, reasonably so, and the sidebar wasn't playing nicely.)

Absolute positioning works well here because the containers are
relatively positioned, so the coords are absolute to the container not
the page.