fix links on release pages not using /pod/$module
modules is an arrayref that will commonly only have one element.
Flattening it to a single value is incorrect.  The code that determines
how to link modules is expecting an arrayref.

Removing the flattenning doesn't seem to cause any issues, and the
hashref that results doesn't have any arrayrefs where it should have
single values.