make remove button on email and website work
The removeLabel JS function seems to not be used at all any more. It operates on elements with the .control class, but that also seems not to be used at all. There are some references to that class in root/static/openlayers/OpenLayers.js, but I am not sure if that's related. Both the JS function as well as the corresponding less code in root/less/account.less can possibly be removed.