make sure all buttons never behabve like submit
Some browsers apparently think that a button is a submit-like button by default. If there's JS on that button, it gets executed when one presses enter in the form. It seems to pick the first button to occur. This only happens in some browsers, and lead to additional email rows being added whenever enter was pressed in any form field. Also see http://stackoverflow.com/a/13534643/1331451.

This changes the behavior insofar that now whenever enter is pressed, the form is submitted and the profile gets saved.

Fixes #1847