fix login for non-pause users
Logging in as a user that didn't have a PAUSE account connected was
working, but the code to fetch the favorites list (which controls the
login menu) wasn't interpreting that correctly.  The favorites code was
checking for a "profile", which only exists for PAUSE users.

Instead, use the normal authentication routines and the resulting user
object.