move release info collection into ReleaseInfo model
This merges the ReleaseInfo role's logic into the ReleaseInfo model, and
allows it to give all of the relevant release info at once.  This
removes the need to have goofy shared functions to massage the data, or
insert it into the stash.

It also allows the parallelism of the release info requests to be
controlled easier.