use forwarded headers
Use X-Forwarded-For etc headers to override standard headers.  Existing
Plack middleware don't quite cover the use case we have.  ReverseProxy
will only take the nearest link from X-Forwarded-For, which for us will
be the Fastly servers.  Fastly is configured to override any incoming
X-Forwarded-For header, so we can safely use the farthest link, which
will be the end user.

Other middleware don't support the other headers we would like to use.