fix up river layout
Various tweaks to the CSS and markup to make the river gauge layout more
consistent across browsers.

Also removes some needless extra divs in the breadcrumb that served no
purpose.  They were styled as inline, so they effectively functioned as
spans, which without any extra styles apply effectively means they were
no-ops.  There was a bit of javascript relying on the "inline" class
that was applied surrounding the favorite button, but it was just
reusing classes that happened to be there rather than anything
semantically tied to the button.