only use gravatars based on user emails
This removes the use of avatars customizable through the account page.
Instead, gravatars based on the users emails is used, falling back to
identicons.

The customizable avatars did not work well for the multiple sizes we
wanted to display them in.  This was also a problem for our fallback
icon.  The user images would often be served through http, not https,
giving mixed mode alerts.  If the user modified their profile, the
generated gravatar URL would be baked in, preventing updates to their
email address from updating the avatar.