use search/web/v2 api
The search/web/v2 API has a different form.  Rather than each result
being an arrayref, the result is an object with hits containing the top
sub-results.  The total number of sub-results is then available on the
result object.  This gives a more accurate number of sub-results,
without having to return every hit and then throwing away most of them.