Display "Development release" to side bar (under "Module version") with matching text color to the release text in bread brumbs area for dev ("maturity": "developer") releases. This is to make it more clear that a release (especially when the version only has an underscore, and no "TRIAL" or other such clues) is in fact a development release.
When there is only one version for a distribution, display a
disabled drop down with "No other versions" text in place of,
and matching the style of, the "Jump to version" drop down, in
the side bar.

Hide "Latest version" link in side bar when there are no stable
("maturity": "released") releases, which would result in a 404
if followed.