fix sortable tables breaking all page js when given bad sort data
The column used to sort with tablesorter can be taken from a number of
places, some of which are under user control.  These can be invalid,
such as referencing columns that aren't sortable.  Add extra validation
that the sortable configuration is valid before handing it off to
tablesorter.

tablesorter also has built in parsing of the current URL, with no way to
override this.  This is a pretty awful API, so just remove it from the
tablesorter code.  The URL parsing can be handled by our own code, with
proper error handling and priorities.