clean up configuration of API urls
There is no need to distinguish between 'secure' and 'unsecure' API
access, only public and private.  Adjust the methods and configuration
to have just a 'api' and 'api_public' url.

These can be configured directly on the API model or HTML view, but with
fallback to the values configured for the app.