fix showing Catalyst error pages on missing source files
Show now the not found page rather than an internal error page if a
source file is not found.  We could throw to an internal error page on
500 errors, but that seems excessive, and 404 will be accurate for some
of those cases anyway.

Checking the code is necessary because some url forms will generate
internal errors from the back end, which won't be able to be decoded,
resulting in errors and warning.  One example of this is
/source/<%meta.name%>/<%meta.other%>