Revert "no "release" means no canonical"
We should never have pod pages without releases. If these show up in the
logs, we either have bad data we should fix, or need to adjust the code
elsewhere to show a 404. @path may be just the module name, not
including the author/release, so this causing incorrect cache headers.

This reverts commit afb43f333b588d00f7928057e30a46171576947c.