Split notifications up to "build notifications" and "add emails".
Building up a series of Notifications from Permissions is comprised of
two pieces, "create a series of notifications from the Permissions", and
then "add email addresses to the notifications".

Split it up into two separate methods, providing us with the opportunity
to add in other notifications that _aren't_ based on Permissions, while
still leveraging the existing code to add emails into the Notifications.