Show a notification for Modules that have been Deprecated.
Not a Release or Distribution, but a particular Module.

Has different notification verbiage, to help distinguish between entire
Distributions that are Deprecated (and thus the reader should either
find something entirey different, or potentially offer to take over),
and individual Modules that are Deprecated (which likely come in a
Distribution that has an alternative Module already provided).