adjust sidebar width to better match the font size
The font size used to be 13px, with a sidebar size of 175px. The font
size has been changed to 15px, so the sidebar needs to be increased.
Change the sidebar to use a formula for its size based on the font size.

Using approximately the same ratio as the old layout, the new size will be 202px.