try to add mid-ellipses on word boundaries
For release names that are too wide to fit in a table, we try to add
ellipses in the middle. Usually, the start and end of a release name are
the most relevant parts. CSS doesn't support adding ellipses in the
center, so we emulate this by cutting the names near the half way point
and having the first segment be abbreviated.

The previous calculation was to cut at the 3/5 of the string, to be
somewhat near the center but allow for space for the ellipsis. Update
this to search backward from that point for the closest word boundary,
as long as it is further than 1/4 through the string.