fix file listing data in model
Having keys with periods in them is hard to work with, and 'stat' should
be a hash anyway, to be consistent with everywhere else. Fix the format
of the file listing in the model rather than forcing the template to
deal with it.

This should probably be different in the API, but for now adapt it here.