refactor auth to use more normal auth workflow
Rather than having the auth and session handling done in the controller,
move it into auth plugins. This means the normal ->authenticate method
works as a login rather than a session restore, and ->logout works.

Move to a fully implemented user and store class, to avoid the weirdness
of the Proxy plugin creating user objects that don't represent users.