fix changes display collapsing
Previously it was possible for the show/more less button on the changes
block to be displayed even if the hidden content was a single line,
which was rather pointless. And after changing the font size, the
calculated sizes could be off by a fraction of a pixel, resulting in
showing the button but it doing nothing.

Fix collapsing to not apply if javascript is disabled, and include the
size of the button when calculating if it should be collapsable.

Also tweak the collapse height to be an integer multiple of the line
height, and remove some unneeded padding between change groups.