include more transitive deps in cpanfile
Add some additional transitive deps to the cpanfile and snapshot to
cover a wider range of perl versions. The system that is automatically
updating the snapshot is running a different perl version than the
current live servers, so it won't have the same modules available. We
need to ensure all of the potentially required modules are listed in the
snapshot or the production machines won't install them.

Also document this a bit better in the cpanfile by splitting out these
deps to its own section.