fix dependency display with single dependency
The API is not always consistent in what it returns, and array refs with
single entries may be returned as just the item without an array.
Template::Alloy would allow using list functions on non-lists, treating
them as single entry lists. Text::Xslate doesn't do that, so we have to
ensure that the data is in the correct form when coming out of the
model.