Make scrollable area larger for pod pre areas (#2528)
The problem is that `.pod pre` and `.syntaxhighlighter` both apply `overflow` in some way. By turning off `overflow-x` on `.syntaxhighlighter`, we end up with the outer `.pod pre` scroll area, which is larger, since that is the one that applies `padding`. This makes it easier to scroll horizontally on one-line code blocks on mobile.

I noticed this on https://metacpan.org/dist/App-perlimports/view/script/perlimports, it has a lot of examples in the SYNOPSIS that are hard for me with my fat fingers to scroll on my mobile.