fix links generated by syntax highlighter
This updates the URLs generated by the syntax highlighter to match the
URLs used by the test of the site after PR#2512.

It also fixes how it recognises if a source or pod page is being
displayed. Previously it was using the URL of the current page, but it
will now be based on recognising the full page source view, the same as
is used to show line numbers and other features.