clean up GitHub repository URLs
Authors will use various forms of URLs when including repository
metadata. For known git hosts (only GitHub so far) we can clean up those
URLs to be presented better in the UI.

One example of this is including just a clone URL using git://. This
isn't friendly to present in the UI to start with, since it doesn't lead
to a web page, but also GitHub no longer supports git:// cloning. There
is also little purpose of presenting a https web link with a second link
for cloning that includes .git at the end. Either one could be used for
cloning.