cope with authors with nothing visible in their name
In an author has an "empty" name set, various parts of the UI will not
have anything to click on, breaking navigation.

Detect this by checking for word characters in the author name, and
falling back to the PAUSE id if there are none. If the API were to
provide a display_name, either through a similar check like we are using
or as a field that can be filled in, it would be preferred over either
the name or the PAUSE id.