preload account data when possible
Rewrite the user data fetching using modern APIs to avoid the
X-Requested-With header, and add a preload tag to load the login status
data as early as possible. This should decrease the time before we're
able to display the user's account icon.

Also remove the extraneous function when loading user data. It includes
a comment referencing possibly optimizing for when there is no login
cookie. But the login cookie is http only, so we can't check it in
javascript. The response from login_status without a cookie is cached,
so there isn't any real opportunity for further optimization or caching.
So the comment can also be removed as extraneous.