fix month in changelog parser calculations
When parsing the date from a change log, the short month was being
matched, and the index of the match was used as the month. Since the
index starts at 0, this resulted in the month being off by one.

Add one to the index before using it in our short month to number hash.