add role for cleaning up file data
The MIME type from the API is not currently reliable, so we have used
the file name and extension in various places to check the file type.
Add a role which can assist cleaning up the MIME type in one location,
and we can then use it reliably everywhere else.