refactor ReleaseInfo to specify release first and requested data later
A substantial refactoring of ReleaseInfo to allow creating a requesting
object, then later deciding what information to request.

The existing APIs are maintained for now, but in the future users will
be adapted to use the ->via_dist and ->via_release methods. The
full_details option will go away, with the extra details being attached
by the release controller. This ReleaseInfo object can be placed in the
stash immediately, rather than needing the pass the dist, author, or
release information on.

The Orchestrator object coordinates fetching in two batches, info based
on the dist and info based on the release. The order of these depends on
which information you start with. Additional information can be
requested using the ->with_* methods before calling fetch.