Log4Perl MDC: add additional headers
Add some additional headers to the Log4Perl MDC, which can be used for
additional filtering.

For example, filters can be defined like:

    log4perl.filter.DocOnly = Log::Log4perl::Filter::MDC
    log4perl.filter.DocOnly.KeyToMatch = Sec-Fetch-Dest
    log4perl.filter.DocOnly.RegexToMatch = document

    log4perl.filter.IsBrowser = Log::Log4perl::Filter::MDC
    log4perl.filter.IsBrowser.KeyToMatch = Sec-Fetch-Dest
    log4perl.filter.IsBrowser.RegexToMatch = .

    log4perl.filter.OnlyDocFromBrowser = Log::Log4perl::Filter::Boolean
    log4perl.filter.OnlyDocFromBrowser.logic = DocOnly || !IsBrowser

The OnlyDocFromBrowser filter can then be added to an appender:

    log4perl.appender.OUTPUT.Filter = OnlyDocFromBrowser

This would then silence any logging from requests for things like images
or XHR requests, which can help when developing.