Fix handling of non-UTF8 source code
When Text::Xslate is given data with the utf8 flag off, but with valid
non-strict utf8 sequences in it, those sequences will be decoded. That
will then produce errors when attempting to encode it for output. We
only really run into this problem when showing the source of a file,
since those may have arbitrary bytes.

Force the source content given to Xslate to always have the utf8 flag on,
so that Text::Xslate inserts the non-decoded "characters" directly into
the output. They will essentially be mojibake, but this is acceptable
since the original file had badly encoded data in it.