clean up code for pages with no sidebar, and grid layout
Create a mixin template that can be used to remove everything connected
with the sidebar. This removes the actual content from the pages, making
it unnecessary to remove the elements with CSS. A class is still
provided to allow changing the grid layout.

Normalize the home page to mostly work like other no sidebar pages. Also
have it remove the extra elements it does not want, rather than hiding
them. With this cleanup, the separate banner code for the home page is
not needed.

Fix the grid row sizing to use max-content. This fixes various places
that would show large empty spaces for elements that were missing.