Page Layout: revise for consistent primary grid
Update the layout to use a consistent grid by always including all grid
elements. This makes the use of grid-template-columns and
grid-template-rows consistent. We can then use 'auto' for the content
section, which will expand as needed. The sidebar can use fit-content,
which means it will automatically disappear when it has no content. This
eliminates the need for any special class for pages without a sidebar.