test classes more accurately in tests
Change the tests to consistently use contains-token for testing classes
in tests, to match CSS style class matching. This makes the tests more
flexible as minor changes to classes for visual reasons won't prevent
the tests from finding the appropriate content.

Also fixes the contains-token function to correctly return a boolean
value and operate on the correct types.