rewrite changes handling for rss feed
The previous code would batch requests in the controller, manually
generate URLs in the model, and had odd inputs and outputs.

Rewrite the changes model by_release method to accept a simple list of
release paths, handling batching internally, and return a simple
hashref. Doing the batching internally makes it easy to add concurrency
to the requests.

Also fix the version regex to understand v prefixes.