don't always sort changelog entries
Some changelogs have unusual orders, or have made mistakes with
versioning. But most change logs will have the most recent releases at
the beginning or the end of the file. If the first or last release found
matches the version we are looking for, just use the order of the file
as is. Only sort the entries if it isn't found at the start or end.