move documented name and associated module and pod into API
Rather than calculate the documented name, associated module, and
associated pod file in the controller, do it in the API.

Currently, this relies on the requested pod file to calculate, which it
probably shouldn't. But preserve the behavior and pass along the
requested name to be used in the calculation.