split table of contents from pod content
Split the index from the rest of the pod content, allowing it to be
templated differently. This allows its markup to be included with the
static content, rather than being assembled in javascript.

This should make it easier to format it differently in the future, and
should also fix the scroll position changing on page load.