docker compose setup for development
Adds a docker compose configuration for development.

Rather than using uwsgi like the production setup, it runs an nginx
instance proxying to a plackup server. plackup runs as a single process
by default, which we want for more readable debugging output, and lets
us specify files to watch to do automatic reloads. nginx is used to
server the static files.

It is possible to configure this to use uwsgi as well, if desired.

The development server can be run using:
docker compose up --build