GitHub Action to update manifests
When a pull request changes, or merges, checkout the metacpan-k8s
repository and set the image tag used for the deployment of the web app.

Manifests updates during pull request creation or changes ensures that
the process continues to work. When the change lands on the trunk branch
of the repository is the update sent to the metacpan-k8s repository.

The GitHub Application for metacpan is able to bypass the pull request
and commit to main, in the metacpan-k8s repository setting.