syntaxhighlighter from npm
syntaxhighlighter is not entirely designed to be used from npm, so it
requires some weird bits.

For one, the brushes are separate packages with terrible names. The
themes are also badly named, and not correctly packaged.

The theme is written in scss and doesn't have a compiled version
available via npm. This requires adding an scss compiler. In the future,
it would be better to convert all of our code to scss, so this isn't
terrible.