fix syntaxhighlighter styles
The syntax highlighter theme is rather terrible, trying to do its own
css reset with !important on everything. This makes it awful to do our
own overrides of its behavior. It also tries to do things that are no
longer relevant.

Vendor the parts of the syntax highlighter theme that we need, removing
the !important tags. This allows our own code to be simplified as well.