rewrite static middleware to correctly set headers
Surrogate-Control headers were being set for 404 responses for static
files. This would cause Fastly to cache them indefinitely. During a
deploy, there would be a period of time where new servers would link to
js and css assets, which old servers would respond with 404s and get
cached.

Fix the static file serving code to only set the Surrogate-Control
headers for 2xx responses to prevent this. This ended up being a
significant rewrite of the middleware.