add healthcheck end point
Add a /healthcheck end point, returning a simple JSON structure. Use a
template rather than normal JSON encoding to ensure that the template
engine is working correctly.

In the future, this could be extended to check that the API is
available, but for now a simple check that just the web server is
running is good enough.