use temp directory for xslate cache
We were already deleting the xslate cache directory on startup, because
it was having problems with refreshing the templates properly. And in a
deployment, we don't have any permanent storage anyway. So we can just
use a temp directory for the cache.

Since we aren't caching templates anymore, we no longer need the var
directory.