work around syntaxhighlighter bug causing empty blocks on perlsecret
perlsecret.pod has some code blocks that are very short. The syntax
highlighter has a bug when it attempts to strip <![CDATA[ ]]> blocks,
which ends up entirely erasing short code block content. The CDATA
stripping is only enabled when the useScriptTags option is enabled,
which we don't need or want. Work around the bug by disabling
useScriptTags.

This bug had previously been fixed by patching the vendor library to fix
the CDATA stripping. That change was lost in the conversion to
node_modules.